fix for CachedBehavior. [#38 state:resolved responsible:fahad19 milestone:'1.3.0']

+++ models/behaviors/cached.php
@@ -54,8 +54,10 @@ class CachedBehavior extends ModelBehavior {
     function _deleteCachedFiles(&$model) {
         foreach ($this->settings[$model->alias]['prefix'] AS $prefix) {
             $files = glob(TMP.'cache/queries/cake_'.$prefix.'*');
-            foreach ($files AS $file) {
-                unlink($file);
+            if (is_array($files) && count($files) > 0) {
+                foreach ($files AS $file) {
+                    unlink($file);
+                }
             }
         }
     }
